edditional development

+++ CCC.Data.Monitoring.Concrete/Entities/MonitorData.cs
@@ -5,8 +5,8 @@ namespace CCC.Data.Monitoring.Concrete.Entities
     public class MonitorData : ITable
     {
         public int ID { get; set; }
-        public int TalkTime { get; set; }
-        public int AfterCallWorkTime { get; set; }
+        public double TalkTime { get; set; }
+        public double AfterCallWorkTime { get; set; }
         public int Handled { get; set; }
         public int Offered { get; set; }
         public int HandledWithinSL { get; set; }
